refactor: 💡Add createdAt and updatedAt to frontend interfaces
