+++ libs/interfaces/src/lib/index.ts
@@ -5,3 +5,4 @@ export * from './order';
 export * from './pricing';
 export * from './transaction-item';
 export * from './user';
+export * from './util';
