+++ libs/interfaces/src/lib/transaction-item/transaction-item.interface.ts
@@ -1,6 +1,6 @@
-import { DesignDimensionEnum, IDesign, IFamilyTree } from '../design';
+import { DesignDimensionEnum, IBase, IDesign, IFamilyTree } from '..';
 
-export interface ITransactionItem {
+export interface ITransactionItem extends IBase {
   transactionItemId: string;
   orderId: string;
   dimension: DesignDimensionEnum;
