+++ libs/interfaces/src/lib/user/user.interface.ts
@@ -1,3 +1,5 @@
+import { IBase } from '..';
+
 export interface IAuthUser {
   userId: string;
   email: string;
@@ -5,7 +7,7 @@ export interface IAuthUser {
   accessToken: string;
   tokenType: string;
 }
-export interface IUser {
+export interface IUser extends IBase {
   userId: string;
   email: string;
   roles: string[];
