+++ libs/interfaces/src/lib/util/base.interface.ts
@@ -1,0 +1,4 @@
+export interface IBase {
+  createdAt?: Date;
+  updatedAt?: Date;
+}
